Prepare for trouble, and make the second in Python
Another RNG offering from the Misc

The script is self-explanatory, iterate over a list of common ports and bombard every IP. It prints out the progress and result to console.
It hold a special place in the Misc folder because it was a "I see your value now" moment for me and Python. 
It was also somewhat relaxing to just run it and stare once in a while.

The timeout is definitely wrong, though.